MarkCommand: Extract from ChatBoxException to separate class
MarkCommand is currently defined as a nested static class inside
ChatBoxException.

Defining logic-heavy commands inside exception classes violates the
Single Responsibility Principle. It makes the code harder to navigate
and maintain as the project grows.

Let's,
* create a new standalone MarkCommand class in the commands package
* remove the nested implementation from ChatBoxException.java
* update the Parser to import and instantiate the new standalone class

This refactoring improves code organization and ensures each class
has a single, clear purpose.
